JWT authentication and modification in other models to accept the authenticated user

+++ app/Http/Controllers/Api/ApiConsultationController.php
@@ -7,8 +7,9 @@ use App\Http\Requests\ConsultationRequest;
 use App\Http\Requests;
 use App\Patient;
 use App\Consultation;
-use App\User;
 use Api\Formatters\ConsultationTransformer;
+use JWTAuth;
+use Gate;
 
 class ApiConsultationController extends ApiController
 {
@@ -22,7 +23,7 @@ class ApiConsultationController extends ApiController
      */
     public function __construct(ConsultationTransformer $transformer){
         $this->consultationTransformer = $transformer;
-        $this->middleware('auth.basic');
+        // $this->middleware('auth.basic');
     }
 
     /**
@@ -44,10 +45,13 @@ class ApiConsultationController extends ApiController
      * @return \Illuminate\Http\Response
      */
     public function store(Patient $patient, ConsultationRequest $request){
+        $user = JWTAuth::parseToken()->authenticate();
+        if(Gate::denies('create_consultation')){
+            return $this->respondForbidden('Usted no tiene permisos para crear consultas medicas');
+        }
         $consultation = new Consultation($request->all());
-        $consultation->user_id = 1; // HARDCODED
-        $patient->consultation()->save($consultation);
-
+        $consultation->patient_id = $patient_id;
+        $user->consultations()->save($consultation);
         return $this->respondCreated('Consulta médica creada correctamente!');
     }
 
@@ -74,9 +78,18 @@ class ApiConsultationController extends ApiController
      * @return \Illuminate\Http\Response
      */
     public function update(Patient $patient, Consultation $consultation, ConsultationRequest $request){
+        $user = JWTAuth::parseToken()->authenticate();
         if($patient->id == $consultation->patient_id){
-            $consultation->update($request->all());
-            return $this->respondEdited('Consulta médica actualizada correctamente');
+            if(Gate::denies('update_consultation', $consultation)){
+                return $this->respondForbidden('Usted no tiene permisos de actualizar la informacion de esta consulta medica');
+            }
+            else{
+                $consultation->update($request->all());
+                return $this->respondEdited('Consulta médica actualizada correctamente');
+            }
+        }
+        else{
+            return $this->respondNotFound('El paciente que busca no tiene la consulta médica solicitada');
         }
     }
 
@@ -88,8 +101,16 @@ class ApiConsultationController extends ApiController
      */
     public function destroy(Patient $patient, Consultation $consultation){
         if($patient->id == $consultation->patient_id){
-            $consultation->delete();
-            return $this->respondDeleted('Consulta médica borrada correctamente');
+            if(Gate::allows('delete_consultation', $consultation)){
+                $consultation->delete();
+                return $this->respondDeleted('Consulta médica borrada correctamente');
+            }
+            else{
+                return $this->respondNotFound('El paciente que busca no tiene la consulta médica solicitada');
+            }
+        }
+        else{
+            return $this->respondNotFound('El paciente que busca no tiene la consulta médica solicitada');
         }
     }
 }
